requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.2.6
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
